removed ISBN change feature in edit book + UI
package name change: viewMyBook to myBook
hardcoded image block for the cover
ISBN is being displayed for edit book but not editable, scan isbn is removed
removed data and description of the book as it is not being used
