+++ doc/JavaDoc/allclasses-frame.html
@@ -13,7 +13,7 @@
 <div class="indexContainer">
 <ul>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/exchange/AcceptedFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.exchange" target="classFrame">AcceptedFragment</a></li>
-<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook" target="classFrame">AddNewBookActivity</a></li>
+<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook" target="classFrame">AddNewBookActivity</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/library/AllTabFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.library" target="classFrame">AllTabFragment</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/library/AvailableFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.library" target="classFrame">AvailableFragment</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/BlockedUsersActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity" target="classFrame">BlockedUsersActivity</a></li>
@@ -34,7 +34,7 @@
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoomItem.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom" target="classFrame">ChatRoomItem</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/ChatRoomTests.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria" target="classFrame">ChatRoomTests</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/model/holder/ChatViewHolder.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.holder" target="classFrame">ChatViewHolder</a></li>
-<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook" target="classFrame">EditBookActivity</a></li>
+<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook" target="classFrame">EditBookActivity</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/EditMyProfileActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity" target="classFrame">EditMyProfileActivity</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/exchange/ExchangeFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.exchange" target="classFrame">ExchangeFragment</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/controller/ImageController.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.controller" target="classFrame">ImageController</a></li>
