+++ doc/JavaDoc/allclasses-noframe.html
@@ -13,7 +13,7 @@
 <div class="indexContainer">
 <ul>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/exchange/AcceptedFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.exchange">AcceptedFragment</a></li>
-<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></li>
+<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/library/AllTabFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.library">AllTabFragment</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/library/AvailableFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.library">AvailableFragment</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/BlockedUsersActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity">BlockedUsersActivity</a></li>
@@ -34,7 +34,7 @@
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/model/chatroom/ChatRoomItem.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.chatroom">ChatRoomItem</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/ChatRoomTests.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria">ChatRoomTests</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/model/holder/ChatViewHolder.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.holder">ChatViewHolder</a></li>
-<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></li>
+<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/EditMyProfileActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity">EditMyProfileActivity</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/exchange/ExchangeFragment.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.fragment.exchange">ExchangeFragment</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/controller/ImageController.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.controller">ImageController</a></li>
