+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html
@@ -49,7 +49,7 @@ var activeTableTab = "activeTableTab";
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../../../../index.html?ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" target="_top">Frames</a></li>
@@ -91,7 +91,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</div>
+<div class="subTitle">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</div>
 <h2 title="Class AddNewBookActivity" class="title">Class AddNewBookActivity</h2>
 </div>
 <div class="contentContainer">
@@ -120,7 +120,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.AddNewBookActivity</li>
+<li>ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook.AddNewBookActivity</li>
 </ul>
 </li>
 </ul>
@@ -147,7 +147,7 @@ var activeTableTab = "activeTableTab";
 </dl>
 <dl>
 <dt>Direct Known Subclasses:</dt>
-<dd><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></dd>
+<dd><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></dd>
 </dl>
 <hr>
 <br>
@@ -487,7 +487,7 @@ extends android.support.v7.app.AppCompatActivity</pre>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../../../../index.html?ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" target="_top">Frames</a></li>
