+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html
@@ -48,7 +48,7 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
@@ -91,7 +91,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</div>
+<div class="subTitle">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</div>
 <h2 title="Class EditBookActivity" class="title">Class EditBookActivity</h2>
 </div>
 <div class="contentContainer">
@@ -120,10 +120,10 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.AddNewBookActivity</a></li>
+<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook.AddNewBookActivity</a></li>
 <li>
 <ul class="inheritance">
-<li>ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.EditBookActivity</li>
+<li>ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook.EditBookActivity</li>
 </ul>
 </li>
 </ul>
@@ -153,7 +153,7 @@ var activeTableTab = "activeTableTab";
 <hr>
 <br>
 <pre>public class <span class="typeNameLabel">EditBookActivity</span>
-extends <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></pre>
+extends <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></pre>
 <div class="block">The Edit book activity.</div>
 </li>
 </ul>
@@ -252,10 +252,10 @@ extends <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activi
 </tr>
 </table>
 <ul class="blockList">
-<li class="blockList"><a name="methods.inherited.from.class.ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.AddNewBookActivity">
+<li class="blockList"><a name="methods.inherited.from.class.ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook.AddNewBookActivity">
 <!--   -->
 </a>
-<h3>Methods inherited from class&nbsp;ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></h3>
+<h3>Methods inherited from class&nbsp;ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></h3>
 <code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addPhoto--">addPhoto</a>, <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#scanISBN--">scanISBN</a>, <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateAuthor-java.lang.String-">validateAuthor</a>, <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateISBN-java.lang.String-">validateISBN</a>, <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateTitle-java.lang.String-">validateTitle</a></code></li>
 </ul>
 <ul class="blockList">
@@ -361,7 +361,7 @@ extends <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activi
 <pre>protected&nbsp;void&nbsp;onCreate(android.os.Bundle&nbsp;savedInstanceState)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Overrides:</span></dt>
-<dd><code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#onCreate-android.os.Bundle-">onCreate</a></code>&nbsp;in class&nbsp;<code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></code></dd>
+<dd><code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#onCreate-android.os.Bundle-">onCreate</a></code>&nbsp;in class&nbsp;<code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></code></dd>
 </dl>
 </li>
 </ul>
@@ -376,7 +376,7 @@ extends <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activi
 <div class="block">Add book.</div>
 <dl>
 <dt><span class="overrideSpecifyLabel">Overrides:</span></dt>
-<dd><code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addBook--">addBook</a></code>&nbsp;in class&nbsp;<code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></code></dd>
+<dd><code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addBook--">addBook</a></code>&nbsp;in class&nbsp;<code><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></code></dd>
 </dl>
 </li>
 </ul>
@@ -417,7 +417,7 @@ extends <a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activi
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
