+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-frame.html
@@ -3,18 +3,18 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Mon Mar 11 13:09:56 MDT 2019 -->
-<title>ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</title>
+<title>ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</title>
 <meta name="date" content="2019-03-11">
 <link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../../script.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html" target="classFrame">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a></h1>
+<h1 class="bar"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html" target="classFrame">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook" target="classFrame">AddNewBookActivity</a></li>
-<li><a href="EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook" target="classFrame">EditBookActivity</a></li>
+<li><a href="AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook" target="classFrame">AddNewBookActivity</a></li>
+<li><a href="EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook" target="classFrame">EditBookActivity</a></li>
 </ul>
 </div>
 </body>
