+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Mon Mar 11 13:09:56 MDT 2019 -->
-<title>ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</title>
+<title>ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</title>
 <meta name="date" content="2019-03-11">
 <link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../../script.js"></script>
@@ -69,7 +69,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</h1>
+<h1 title="Package" class="title">Package&nbsp;ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</h1>
 </div>
 <div class="contentContainer">
 <ul class="blockList">
@@ -82,13 +82,13 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></td>
+<td class="colFirst"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></td>
 <td class="colLast">
 <div class="block">The Add new book activity.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></td>
+<td class="colFirst"><a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></td>
 <td class="colLast">
 <div class="block">The Edit book activity.</div>
 </td>
