+++ doc/JavaDoc/ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_91) on Mon Mar 11 13:09:56 MDT 2019 -->
-<title>ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook Class Hierarchy</title>
+<title>ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook Class Hierarchy</title>
 <meta name="date" content="2019-03-11">
 <link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../../script.js"></script>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook Class Hierarchy";
+            parent.document.title="ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook Class Hierarchy";
         }
     }
     catch(err) {
@@ -69,7 +69,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 class="title">Hierarchy For Package ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</h1>
+<h1 class="title">Hierarchy For Package ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../../../../../overview-tree.html">All Packages</a></li>
@@ -94,9 +94,9 @@
 <ul>
 <li type="circle">android.support.v7.app.AppCompatActivity (implements android.support.v7.app.ActionBarDrawerToggle.DelegateProvider, android.support.v7.app.AppCompatCallback, android.support.v4.app.TaskStackBuilder.SupportParentable)
 <ul>
-<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">AddNewBookActivity</span></a>
+<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">AddNewBookActivity</span></a>
 <ul>
-<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">EditBookActivity</span></a></li>
+<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../../../../../../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">EditBookActivity</span></a></li>
 </ul>
 </li>
 </ul>
