+++ doc/JavaDoc/index-files/index-1.html
@@ -94,11 +94,11 @@
 </dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/UserTests.html#addBlockedUsersTest--">addBlockedUsersTest()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/UserTests.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria">UserTests</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addBook--">addBook()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addBook--">addBook()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>
 <div class="block">Add book.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#addBook--">addBook()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#addBook--">addBook()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/controller/BookController.html#addBook-ca.ualberta.CMPUT3012019T02.alexandria.model.Book-">addBook(Book)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.controller.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/controller/BookController.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.controller">BookController</a></dt>
 <dd>
@@ -141,11 +141,11 @@
 <dd>
 <div class="block">Adds an owned book to the current user's collection of owned books</div>
 </dd>
-<dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">AddNewBookActivity</span></a> - Class in <a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a></dt>
+<dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">AddNewBookActivity</span></a> - Class in <a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a></dt>
 <dd>
 <div class="block">The Add new book activity.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#AddNewBookActivity--">AddNewBookActivity()</a></span> - Constructor for class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#AddNewBookActivity--">AddNewBookActivity()</a></span> - Constructor for class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/user/User.html#addOwnedBook-ca.ualberta.CMPUT3012019T02.alexandria.model.user.OwnedBook-">addOwnedBook(OwnedBook)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.model.user.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/model/user/User.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.model.user">User</a></dt>
 <dd>
@@ -158,7 +158,7 @@
 <dd>
 <div class="block">Adds photo from camera</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addPhoto--">addPhoto()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#addPhoto--">addPhoto()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>
 <div class="block">Add photo.</div>
 </dd>
