+++ doc/JavaDoc/index-files/index-11.html
@@ -107,9 +107,9 @@
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/SignUpActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/SignUpActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity">SignUpActivity</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/ViewUserProfileActivity.html#onCreate-android.os.Bundle-">onCreate(Bundle)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/ViewUserProfileActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity">ViewUserProfileActivity</a></dt>
 <dd>&nbsp;</dd>
