+++ doc/JavaDoc/index-files/index-13.html
@@ -73,7 +73,7 @@
 </a>
 <h2 class="title">S</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#scanISBN--">scanISBN()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#scanISBN--">scanISBN()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>
 <div class="block">Scan isbn.</div>
 </dd>
