+++ doc/JavaDoc/index-files/index-16.html
@@ -73,15 +73,15 @@
 </a>
 <h2 class="title">V</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateAuthor-java.lang.String-">validateAuthor(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateAuthor-java.lang.String-">validateAuthor(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>
 <div class="block">Validate author boolean.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateISBN-java.lang.String-">validateISBN(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateISBN-java.lang.String-">validateISBN(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>
 <div class="block">Validate isbn boolean.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateTitle-java.lang.String-">validateTitle(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">AddNewBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html#validateTitle-java.lang.String-">validateTitle(String)</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">AddNewBookActivity</a></dt>
 <dd>
 <div class="block">Validate title boolean.</div>
 </dd>
