+++ doc/JavaDoc/index-files/index-3.html
@@ -77,7 +77,7 @@
 <dd>&nbsp;</dd>
 <dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity</a> - package ca.ualberta.CMPUT3012019T02.alexandria.activity</dt>
 <dd>&nbsp;</dd>
-<dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a> - package ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</dt>
+<dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a> - package ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</dt>
 <dd>&nbsp;</dd>
 <dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/adapter/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.adapter</a> - package ca.ualberta.CMPUT3012019T02.alexandria.adapter</dt>
 <dd>&nbsp;</dd>
