+++ doc/JavaDoc/index-files/index-4.html
@@ -81,7 +81,7 @@
 <dd>
 <div class="block">As the current user (book owner), decline a request on your book made by another user</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#deleteBook--">deleteBook()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#deleteBook--">deleteBook()</a></span> - Method in class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></dt>
 <dd>
 <div class="block">Delete book.</div>
 </dd>
