+++ doc/JavaDoc/index-files/index-5.html
@@ -73,11 +73,11 @@
 </a>
 <h2 class="title">E</h2>
 <dl>
-<dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">EditBookActivity</span></a> - Class in <a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a></dt>
+<dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">EditBookActivity</span></a> - Class in <a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a></dt>
 <dd>
 <div class="block">The Edit book activity.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#EditBookActivity--">EditBookActivity()</a></span> - Constructor for class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook">EditBookActivity</a></dt>
+<dt><span class="memberNameLink"><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html#EditBookActivity--">EditBookActivity()</a></span> - Constructor for class ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook">EditBookActivity</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/EditMyProfileActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity"><span class="typeNameLink">EditMyProfileActivity</span></a> - Class in <a href="../ca/ualberta/CMPUT3012019T02/alexandria/activity/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity</a></dt>
 <dd>
