+++ doc/JavaDoc/overview-frame.html
@@ -15,7 +15,7 @@
 <ul title="Packages">
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria.activity</a></li>
-<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a></li>
+<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/adapter/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria.adapter</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/controller/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria.controller</a></li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/package-frame.html" target="packageFrame">ca.ualberta.CMPUT3012019T02.alexandria.fragment</a></li>
