+++ doc/JavaDoc/overview-summary.html
@@ -85,7 +85,7 @@
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a></td>
+<td class="colFirst"><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-summary.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a></td>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
