+++ doc/JavaDoc/overview-tree.html
@@ -74,7 +74,7 @@
 <ul class="horizontal">
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria</a>, </li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria.activity</a>, </li>
-<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook</a>, </li>
+<li><a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook</a>, </li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/adapter/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria.adapter</a>, </li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/controller/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria.controller</a>, </li>
 <li><a href="ca/ualberta/CMPUT3012019T02/alexandria/fragment/package-tree.html">ca.ualberta.CMPUT3012019T02.alexandria.fragment</a>, </li>
@@ -116,9 +116,9 @@
 <ul>
 <li type="circle">android.support.v7.app.AppCompatActivity (implements android.support.v7.app.ActionBarDrawerToggle.DelegateProvider, android.support.v7.app.AppCompatCallback, android.support.v4.app.TaskStackBuilder.SupportParentable)
 <ul>
-<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">AddNewBookActivity</span></a>
+<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/AddNewBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">AddNewBookActivity</span></a>
 <ul>
-<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook"><span class="typeNameLink">EditBookActivity</span></a></li>
+<li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.viewmybook.<a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/viewmybook/EditBookActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity.myBook"><span class="typeNameLink">EditBookActivity</span></a></li>
 </ul>
 </li>
 <li type="circle">ca.ualberta.CMPUT3012019T02.alexandria.activity.<a href="ca/ualberta/CMPUT3012019T02/alexandria/activity/BlockedUsersActivity.html" title="class in ca.ualberta.CMPUT3012019T02.alexandria.activity"><span class="typeNameLink">BlockedUsersActivity</span></a></li>
